Rename test_$package.py files to simply test.py, and move AsyncMock up.
