+++ maas/client/viscera/testing.py
@@ -1,17 +1,15 @@
 """ Testing framework for maas.client.viscera """
 
 __all__ = [
-    'AsyncMock',
-    'Awaitable',
     'bind',
 ]
 
-from functools import partial
 from itertools import chain
 from typing import Mapping
 from unittest.mock import Mock
 
 from . import OriginBase
+from ..testing import AsyncMock
 
 
 def bind(*objects, session=None):
@@ -49,28 +47,3 @@ def bind(*objects, session=None):
         }
 
     return OriginBase(session, objects=objects)
-
-
-class AsyncMock(Mock):
-    """Mock that is "future-like"; see PEP-492 for the details.
-
-    The new `await` syntax chokes on arguments that are not future-like, i.e.
-    have an `__await__` call, so we have to fool it.
-    """
-
-    def __call__(_mock_self, *args, **kwargs):
-        callup = super(AsyncMock, _mock_self).__call__
-        call = partial(callup, *args, **kwargs)
-        return Awaitable(call)
-
-
-class Awaitable:
-    """Wrap a "normal" call in a future-like object."""
-
-    def __init__(self, call):
-        super(Awaitable, self).__init__()
-        self._call = call
-
-    def __await__(self):
-        yield  # This serves only to make this a generator.
-        return self._call()
